# api/openapi.yaml
openapi: 3.0.0
paths:
...
  /orders:
    post:
      operationId: createOrder
      parameters: []
      responses:
        "201":
          description: ""
        "401": &a10
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Missing Authorization Header
        "403": &a11
          description: User has no role assigned or insufficient permissions
          content:
            application/json:
              schema:
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: Insufficient permissions
        "500": &a12
          description: Internal server error
          content:
            application/json:
              schema: *a1
      security: &a13
        - bearer: []
      tags: &a14
        - Orders
    get:
      operationId: searchOrders
      parameters: []
      responses:
        "200":
          description: ""
        "401": *a10
        "403": *a11
        "500": *a12
      security: *a13
      tags: *a14
  /orders/{uuid}:
    get:
      operationId: getOrder
      parameters:
        - name: uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "401": *a10
        "403": *a11
        "500": *a12
      security: *a13
      tags: *a14
    patch:
      operationId: updateStatus
      parameters:
        - name: uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderDto"
      responses:
        "200":
          description: ""
        "401": *a10
        "403": *a11
        "500": *a12
      security: *a13
      tags: *a14
info:
  title: Nexus API
  description: The backend e-commerce API for Nexus store
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateOrderDto:
      type: object
      properties: {}